Logic in GalleryController moved to GalleryViewModel, search no longer requires cookie yet persists, added safeguard in SessionTimeoutController.

+++ Controllers/SessionTimeoutController.cs
@@ -20,11 +20,15 @@ namespace SA51_CA_Project_Team10.Controllers
         {
             string sessionId = HttpContext.Request.Cookies["sessionId"];
 
-            _db.Sessions.Remove(new Session()
+            // Should not be naturally possible to error here, but this safeguards against errors caused by database de-sync during testing
+            if (_db.Sessions.FirstOrDefault(session => session.Id == sessionId) != null)
             {
-                Id = sessionId
-            });
-            _db.SaveChanges();
+                _db.Sessions.Remove(new Session()
+                {
+                    Id = sessionId
+                });
+                _db.SaveChanges();
+            }
 
             HttpContext.Response.Cookies.Delete("sessionId");
